Refactor sidebar components into separate files
