+++ src/pages/(charts)/+Layout.tsx
@@ -1,11 +1,9 @@
-import { FC, ReactNode, useState } from 'react';
+import { FC, ReactNode } from 'react';
 
-import { SimpleSelect } from '@/components/SimpleSelect';
-import { Drawer, DrawerContent } from '@/components/ui/drawer';
 import { ScrollArea } from '@/components/ui/scroll-area';
-import { useBreakpoint } from '@/lib/hooks/tailwind';
 
 import { Levers } from './(_components)/Levers';
+import { ScenariosSidebarLayout } from './(_components)/ScenariosSidebarLayout';
 
 export const Layout: FC<{ children?: ReactNode }> = ({ children }) => {
   return (
@@ -13,10 +11,10 @@ export const Layout: FC<{ children?: ReactNode }> = ({ children }) => {
       {/* Main content */}
       <div className="grow shrink max-h-full flex flex-col items-center">
         <div className="grow shrink w-screen max-w-screen lg:max-w-5xl flex flex-row items-stretch">
-          <ScenariosSidebar>
+          <ScenariosSidebarLayout>
             {/* <Presets /> */}
             <Levers />
-          </ScenariosSidebar>
+          </ScenariosSidebarLayout>
           <div className="grow shrink">
             <ScrollArea className="h-[calc(100vh-56px)]">
               {children}
@@ -28,51 +26,3 @@ export const Layout: FC<{ children?: ReactNode }> = ({ children }) => {
     </>
   );
 };
-
-export function ScenariosSidebar({ children }: { children: React.ReactNode }) {
-  const isDesktop = useBreakpoint('md');
-  return isDesktop ? <DesktopSidebar>{children}</DesktopSidebar> : <MobileDrawer>{children}</MobileDrawer>;
-}
-
-function DesktopSidebar({ children }: { children: ReactNode }) {
-  return (
-    <div className="w-72 shrink-0 border-r-2 border-r-gray-200">
-      <ScrollArea className="h-[calc(100vh-56px)]">
-        <div className="py-10 px-5">{children}</div>
-      </ScrollArea>
-    </div>
-  );
-}
-
-function MobileDrawer({ children }: { children: ReactNode }) {
-  const [activeSnapPoint, setActiveSnapPoint] = useState<string | number | null>(0.5);
-
-  return (
-    <Drawer
-      dismissible={false}
-      modal={false}
-      open={true}
-      snapPoints={['50px', '300px', 0.5]}
-      activeSnapPoint={activeSnapPoint}
-      setActiveSnapPoint={setActiveSnapPoint}
-      shouldScaleBackground={false}
-      noBodyStyles={true}
-      handleOnly={true}
-    >
-      <DrawerContent className="h-screen z-40">
-        <ScrollArea className="h-[calc(100vh-var(--snap-point-height)-60px)] mt-10 pb-10 px-5">{children}</ScrollArea>
-      </DrawerContent>
-    </Drawer>
-  );
-}
-
-const presetOptions = [
-  { value: '11111', label: 'Scenario 1' },
-  { value: '11112', label: 'Scenario 2' },
-  { value: '11113', label: 'Scenario 3' },
-];
-
-function Presets() {
-  const [value, onChange] = useState(presetOptions[0].value);
-  return <SimpleSelect options={presetOptions} value={value} onChange={onChange} />;
-}
